refactor: 리포지토리를 인터페이스와 Jdbc 구현체로 분리
