+++ src/main/java/roomescape/repository/JdbcReservationTimeRepository.java
@@ -1,0 +1,73 @@
+package roomescape.repository;
+
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+import roomescape.domain.ReservationTime;
+
+import java.sql.PreparedStatement;
+import java.time.LocalTime;
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public class JdbcReservationTimeRepository implements ReservationTimeRepository {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public JdbcReservationTimeRepository(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public List<ReservationTime> findAll() {
+        String sql = "SELECT * FROM reservation_time";
+        return jdbcTemplate.query(sql, reservationTimeRowsMapper());
+    }
+
+    @Override
+    public Optional<ReservationTime> findById(long timeId) {
+        String sql = "SELECT * FROM reservation_time WHERE id = ?";
+        try {
+            ReservationTime time = jdbcTemplate.queryForObject(sql, (rs, rowNum) -> {
+                Long id = rs.getLong("id");
+                LocalTime startAt = rs.getObject("start_at", LocalTime.class);
+                return new ReservationTime(id, startAt);
+            }, timeId);
+            return Optional.ofNullable(time);
+        } catch (EmptyResultDataAccessException e) {
+            return Optional.empty();
+        }
+    }
+
+    @Override
+    public ReservationTime save(ReservationTime reservationTime) {
+        String sql = "INSERT INTO reservation_time (start_at) VALUES (?)";
+
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
+            ps.setObject(1, reservationTime.getStartAt());
+            return ps;
+        }, keyHolder);
+
+        long id = keyHolder.getKey().longValue();
+        return new ReservationTime(id, reservationTime.getStartAt());
+    }
+
+    @Override
+    public void deleteById(long id) {
+        String sql = "DELETE FROM reservation_time WHERE id = ?";
+        jdbcTemplate.update(sql, id);
+    }
+
+    private RowMapper<ReservationTime> reservationTimeRowsMapper() {
+        return (rs, rowNum) -> new ReservationTime(
+                rs.getLong("id"),
+                rs.getObject("start_at", LocalTime.class)
+        );
+    }
+}
