+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -1,69 +1,17 @@
 package roomescape.repository;
 
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
 import roomescape.domain.ReservationTime;
 
-import java.sql.PreparedStatement;
-import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
 
-@Repository
-public class ReservationTimeRepository {
+public interface ReservationTimeRepository {
 
-    private final JdbcTemplate jdbcTemplate;
+    List<ReservationTime> findAll();
 
-    public ReservationTimeRepository(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
+    Optional<ReservationTime> findById(long timeId);
 
-    public List<ReservationTime> findAll() {
-        String sql = "SELECT * FROM reservation_time";
-        return jdbcTemplate.query(sql, reservationTimeRowsMapper());
-    }
+    ReservationTime save(ReservationTime reservationTime);
 
-    public Optional<ReservationTime> findById(long timeId) {
-        String sql = "SELECT * FROM reservation_time WHERE id = ?";
-        try {
-            ReservationTime time = jdbcTemplate.queryForObject(sql, (rs, rowNum) -> {
-                Long id = rs.getLong("id");
-                LocalTime startAt = rs.getObject("start_at", LocalTime.class);
-                return new ReservationTime(id, startAt);
-            }, timeId);
-            return Optional.ofNullable(time);
-        } catch (EmptyResultDataAccessException e) {
-            return Optional.empty();
-        }
-    }
-
-    public ReservationTime save(ReservationTime reservationTime) {
-        String sql = "INSERT INTO reservation_time (start_at) VALUES (?)";
-
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        jdbcTemplate.update(connection -> {
-            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
-            ps.setObject(1, reservationTime.getStartAt());
-            return ps;
-        }, keyHolder);
-
-        long id = keyHolder.getKey().longValue();
-        return new ReservationTime(id, reservationTime.getStartAt());
-    }
-
-    public void deleteById(long id) {
-        String sql = "DELETE FROM reservation_time WHERE id = ?";
-        jdbcTemplate.update(sql, id);
-    }
-
-    private RowMapper<ReservationTime> reservationTimeRowsMapper() {
-        return (rs, rowNum) -> new ReservationTime(
-                rs.getLong("id"),
-                rs.getObject("start_at", LocalTime.class)
-        );
-    }
+    void deleteById(long id);
 }
